+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 5.6.0 (Next)
 
 * [#145](https://github.com/radar/distance_of_time_in_words/pull/145): Add `dotiw/core` require path to use dotiw without overriding the Rails date helpers, `DOTIW.distance_of_time_in_words` shortcut - [@ermolaev](https://github.com/ermolaev).
+* [#146](https://github.com/radar/distance_of_time_in_words/pull/146): Add support for Rails 8 - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ## 5.5.0 (2026/03/15)